requirements.txt
streamlit
python-dotenv
langchain
openai
chromadb
sentence-transformers
PyPDF2
tiktoken